requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
pdf2image==1.16.3
Pillow==10.4.0
numpy==1.24.3
gunicorn==21.2.0
python-dotenv==1.0.0